Rewrite Tabs
